+++ nodes/HetznerCloud/v1/actions/image/list/index.ts
@@ -1,4 +1,4 @@
 import { list as execute } from './execute';
-import { snapshotListDescription as description } from './description';
+import { imageListDescription as description } from './description';
 
 export { description, execute };