+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
-## Released (2024-02-02 0.0.1)
+## v0.0.2 (2024-02-02)
+
+Finish the development of three basic functions:
+
+- list servers (with filters)
+- list images
+- create image/snapshot of an existing server
+
+## v0.0.1 (2024-02-01)
 
 Initial version with basic authentication and a first module to interact with snapshots and servers.